added PatrollingGuard, simulation stops when one team wins
I added a ParollingGuard that walks in a circle between three randomly picked cornerpoints (using the same pathfinding the intruders have) of the board. I also made some small changes so that now if the intruders or guards win, the simulation stops and a message indicating who won is printed.